+++ MonoGame.Framework/Platform/Graphics/TextureCube.OpenGL.cs
@@ -163,7 +163,7 @@ namespace Microsoft.Xna.Framework.Graphics
                 // Use try..finally to make sure dataHandle is freed in case of an error
                 try
                 {
-                    var startBytes = startIndex * elementSizeInByte;
+                    var startBytes = (long)startIndex * elementSizeInByte;
                     var dataPtr = new IntPtr(dataHandle.AddrOfPinnedObject().ToInt64() + startBytes);
 
                     GL.BindTexture(TextureTarget.TextureCubeMap, this.glTexture);
